// src/app/common/model/bf-coupon.model.ts
export class BfCoupon {
  id?: any;
  couponCode?: any;
  couponType?: any;
  couponName?: any;
  organizationId?: any;
  distributor?: any;
  money?: any;
  enable?: any;
  effectiveTime?: any;
  chargeCode?: any;
  idt?: any;
  udt?: any;
}
export class AddBfCoupon {
  couponName?: any;
  couponType?: any;
  chargeCode?: any;
  effectiveTime?: any;
  money?: any;
}
export class ModifyBfCoupon {
  id?: any;
  couponCode?: any;
  couponType?: any;
  couponName?: any;
  organizationId?: any;
  distributor?: any;
  money?: any;
  enable?: any;
  effectiveTime?: any;
  chargeCode?: any;
  idt?: any;
  udt?: any;
}